Tell Taskbar to be hidden when we are fullscreen
When a window_manager window is maximized, the shell treats the window
as fullscreen. When Taskbar is set to auto-hide, and the window is
maximized, the taskbar will not pop up.

This sets NonRudeHWND property before showing the window. This requires
a hide/show cycle when we enter or exit the fullscreen mode.

Fixes #438